+++ vllm/model_executor/models/llama.py
@@ -27,10 +27,8 @@ import torch
 from torch import nn
 from transformers import LlamaConfig
 
-import vllm.envs as envs
 from vllm.attention import Attention, AttentionMetadata
-from vllm.compilation import forward_context
-from vllm.compilation.wrapper import TorchCompileWrapperWithCustomDispatcher
+from vllm.compilation.decorators import support_compile_llama_style
 from vllm.config import CacheConfig, LoRAConfig
 from vllm.distributed import (get_pp_group, get_tensor_model_parallel_rank,
                               get_tensor_model_parallel_world_size)
@@ -347,8 +345,8 @@ class LlamaModel(nn.Module):
         return hidden_states
 
 
-class LlamaForCausalLM(nn.Module, SupportsLoRA,
-                       TorchCompileWrapperWithCustomDispatcher):
+@support_compile_llama_style
+class LlamaForCausalLM(nn.Module, SupportsLoRA):
     packed_modules_mapping = {
         "qkv_proj": [
             "q_proj",
@@ -440,41 +438,6 @@ class LlamaForCausalLM(nn.Module, SupportsLoRA,
             self.sampler = Sampler()
         else:
             self.lm_head = PPMissingLayer()
-        self._use_torch_compile = envs.VLLM_TORCH_COMPILE_LEVEL > 0
-        if self._use_torch_compile:
-            TorchCompileWrapperWithCustomDispatcher.__init__(self)
-
-    def need_to_specialize(self, runtime_shapes: Tuple[int, ...]) -> bool:
-        if len(self.sizes_to_specialize) == 0:
-            return False
-        return runtime_shapes[0] in self.sizes_to_specialize
-
-    def __call__(
-        self,
-        input_ids: torch.Tensor,
-        positions: torch.Tensor,
-        kv_caches: List[torch.Tensor],
-        attn_metadata: AttentionMetadata,
-        intermediate_tensors: Optional[IntermediateTensors] = None,
-    ) -> Union[torch.Tensor, IntermediateTensors]:
-        if not self._use_torch_compile:
-            return self.forward(input_ids, positions, kv_caches, attn_metadata,
-                                intermediate_tensors)
-        with forward_context(attn_metadata):
-            if len(self.compiled_codes) < 1:
-                torch._dynamo.mark_dynamic(input_ids, 0)
-                torch._dynamo.mark_dynamic(positions, 0)
-                if intermediate_tensors is not None:
-                    for tensors in intermediate_tensors.tensors.values():
-                        torch._dynamo.mark_dynamic(tensors, 0)
-                return self.compiled_callable(input_ids, positions, kv_caches,
-                                              attn_metadata,
-                                              intermediate_tensors)
-            with self.dispatch_to_code(0):
-                model_output = self.forward(input_ids, positions, kv_caches,
-                                            attn_metadata,
-                                            intermediate_tensors)
-            return model_output
 
     def forward(
         self,
